refactor: create get-entry-relative-to in utils
