+++ lib/components/app/app-menu-item.tsx
@@ -1,9 +1,11 @@
 import { ChevronDown } from '@styled-icons/fa-solid/ChevronDown'
 import { ChevronUp } from '@styled-icons/fa-solid/ChevronUp'
-import AnimateHeight from 'react-animate-height'
-import React, { Component, HTMLAttributes, KeyboardEvent } from 'react'
 
+import { getEntryRelativeTo } from '../util/get-entry-relative-to'
+import AnimateHeight from 'react-animate-height'
 import Link from '../util/link'
+
+import React, { Component, HTMLAttributes, KeyboardEvent } from 'react'
 
 interface Props extends HTMLAttributes<HTMLElement> {
   href?: string
@@ -17,20 +19,8 @@ interface Props extends HTMLAttributes<HTMLElement> {
 interface State {
   isExpanded: boolean
 }
-
-/**
- * Helper method to find the element within the app menu at the given offset
- * (e.g. previous or next) relative to the specified element.
- * The query is limited to the app menu so that arrow navigation is contained within
- * (tab navigation is not restricted).
- */
-function getEntryRelativeTo(element: EventTarget, offset: 1 | -1): HTMLElement {
-  const entries = Array.from(
-    document.querySelectorAll('.app-menu a, .app-menu button')
-  )
-  const elementIndex = entries.indexOf(element as HTMLElement)
-  return entries[elementIndex + offset] as HTMLElement
-}
+// Argument for document.querySelectorAll to target focusable elements.
+const queryId = '.app-menu a, .app-menu button'
 
 /**
  * Renders a single entry from the hamburger menu.
@@ -48,13 +38,13 @@ export default class AppMenuItem extends Component<Props, State> {
         subItems && this.setState({ isExpanded: false })
         break
       case 'ArrowUp':
-        getEntryRelativeTo(element, -1)?.focus()
+        getEntryRelativeTo(queryId, element, -1)?.focus()
         break
       case 'ArrowRight':
         subItems && this.setState({ isExpanded: true })
         break
       case 'ArrowDown':
-        getEntryRelativeTo(element, 1)?.focus()
+        getEntryRelativeTo(queryId, element, 1)?.focus()
         break
       case ' ':
         // For links (tagName "A" uppercase), trigger link on space for consistency with buttons.
